requirements.txt
altair==5.1.2
attrs==23.1.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy @ file:///C:/b/abs_d5rpy288vc/croots/recipe/contourpy_1663827418189/work
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
docopt==0.6.2
et-xmlfile==1.1.0
fonttools==4.25.0
gitdb==4.0.11
GitPython==3.1.40
idna==3.4
importlib-metadata==6.8.0
importlib-resources @ file:///C:/b/abs_9d6rmtszbl/croot/importlib_resources_1698254653090/work
Jinja2==3.1.2
joblib @ file:///C:/b/abs_1anqjntpan/croot/joblib_1685113317150/work
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
kiwisolver @ file:///C:/b/abs_88mdhvtahm/croot/kiwisolver_1672387921783/work
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib @ file:///C:/b/abs_e26vnvd5s1/croot/matplotlib-suite_1698692153288/work
mdurl==0.1.2
mkl-fft @ file:///C:/b/abs_19i1y8ykas/croot/mkl_fft_1695058226480/work
mkl-random @ file:///C:/b/abs_edwkj1_o69/croot/mkl_random_1695059866750/work
mkl-service==2.4.0
munkres==1.1.4
numexpr @ file:///C:/b/abs_5fucrty5dc/croot/numexpr_1696515448831/work
numpy @ file:///C:/b/abs_9fu2cs2527/croot/numpy_and_numpy_base_1695830496596/work/dist/numpy-1.26.0-cp39-cp39-win_amd64.whl#sha256=9df7f35b8a04328650588772e3df359b51f1e83651ab665c98af9dbcf9a088c2
openpyxl==3.0.10
packaging @ file:///C:/b/abs_28t5mcoltc/croot/packaging_1693575224052/work
pandas @ file:///C:/b/abs_81b3npjrl8/croot/pandas_1697477219064/work/dist/pandas-2.1.1-cp39-cp39-win_amd64.whl#sha256=52f4b2edc1ff7d4ab3efc3bc5744e39c3cdbd6d723393d35fe439f4e139fde7a
patsy==0.5.3
Pillow @ file:///C:/b/abs_20ztcm8lgk/croot/pillow_1696580089746/work
pipreqs==0.4.13
protobuf==4.25.1
pyarrow==14.0.1
pydeck==0.8.1b0
Pygments==2.16.1
pyparsing @ file:///C:/Users/BUILDE~1/AppData/Local/Temp/abs_7f_7lba6rl/croots/recipe/pyparsing_1661452540662/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz @ file:///C:/b/abs_19q3ljkez4/croot/pytz_1695131651401/work
referencing==0.31.0
requests==2.31.0
rich==13.7.0
rpds-py==0.13.0
scikit-learn @ file:///C:/b/abs_55olq_4gzc/croot/scikit-learn_1690978955123/work
scipy==1.11.3
seaborn @ file:///C:/b/abs_68ltdkoyoo/croot/seaborn_1673479199997/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.1
statsmodels @ file:///C:/b/abs_7bth810rna/croot/statsmodels_1689937298619/work
streamlit==1.28.2
tenacity==8.2.3
threadpoolctl @ file:///Users/ktietz/demo/mc3/conda-bld/threadpoolctl_1629802263681/work
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
typing_extensions==4.8.0
tzdata @ file:///croot/python-tzdata_1690578112552/work
tzlocal==5.2
urllib3==2.1.0
validators==0.22.0
watchdog==3.0.0
yarg==0.1.9
zipp @ file:///C:/b/abs_b9jfdr908q/croot/zipp_1672387552360/work
